requirements.txt
torch==2.3.1
torchvision==0.18.1
numpy==1.26.4
Pillow==10.3.0
matplotlib==3.9.0
Werkzeug==3.0.3
Flask==3.0.3
requests==2.32.3